tx-asset-cycle/pyproject.toml
[project]
name = "asset-cycle"
version = "0.1.0"
description = ""
authors = [
    {name = "drangga@vizalogix.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3 (==1.38.18)",
    "psycopg2-binary (==2.9.10)",
    "shapely (==2.1.1)",
    "redis (==6.2.0)",
    "amazon-kclpy (==3.0.1)",
    "psutil (==7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
